requirements.txt
torch
torchvision
matplotlib
tqdm
PILLOW
scipy
numpy
torchmetrics 